requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
crispy-bootstrap5==0.6
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-appconf==1.0.5
django-bootstrap-form==3.4
django-bootstrap3==21.2
django-compressor==4.0
django-crispy-bulma==0.2
django-crispy-forms==1.14.0
django-filter==21.1
django-libsass==0.9
django-on-heroku==1.1.2
django-sass-processor==1.1
django-widget-tweaks==1.4.12
django-widgets-improved==1.5.0
environ==1.0
executing==0.8.3
gunicorn==20.1.0
idna==3.3
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
libsass==0.21.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
oauthlib==3.2.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
protobuf==3.20.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
PyMySQL==1.0.2
python3-openid==3.2.0
rcssmin==1.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.0.0
